Move estimate components
